fix: link with @loader_path/libffmpeg.dylib
Submitted to https://chromium-review.googlesource.com/c/chromium/third_party/ffmpeg/+/3803946.

When building with `is_component_build=false is_component_ffmpeg=true`,
we must manually instruct executables to link with the ffmpeg.dylib,
which is generated at the @loader_path, where for most targets is the
out/{Release,Debug} dir.

Using @rpath is wrong because the @rpath of most targets does not
include the out dir, and the linker won't be able to find ffmpeg.dylib
because of so.

Patch-Filename: link_with_loader_path.patch